Provide IO stats for the pool and disks; provide histogram sums
This is two changes that were developed together and have become
intertwined.

This extends IO stats from being just for vdevs to also providing them
for pools and disks, depending on the setting of the new -depth
command line option. Not all IO stats are meaningful for all levels,
but we do what we can and we report what's available.

We provide sums for histograms in addition to the previous bucket
counts and overall count. ZFS itself doesn't provide a total sum, so
we have to make it up in the same way that 'zpool iostat' does, by
assuming that every bucket value is at its midpoint. Providing sums
makes histograms more useful in practice in various environments, and
you can't otherwise calculate them in eg Prometheus's PromQL (a
histogram quantile estimate isn't the same thing).