Improve vdev naming to match OpenZFS 'zpool status' output
We now match what OpenZFS reports as the names of raidz and draid
vdevs, at least in simple tests. Draid vdevs have intricate names and
reporting them in their full form may be a mistake if they can ever be
reshaped. Also possibly we should report the additional component
information for draid vdevs in the way that we already report
"nparity" if it's available for a vdev.

(A reshaped draid vdev would change the vdev="" label value, which will
create a new set of time series for that vdev which are distinct from
the time series before the reshaping. But perhaps that would be okay,
since the vdev would likely have different performance characteristics.)